PROJECT RESET
The project is being reset as I've resolved to make the game for VR, but some bloat in code is also to be eliminated.

Part of my dream in becoming a game developer was to make VR games, but I became discouraged based on relative market size compared to standard screen titles. Considering that I am making this on my own and I have no specific expectation for revenue in the first place, I will resolve to make the game I want to make.